Add notification type to message, little cleanup
icinga2 has a "notification type". This is usually
"PROBLEM" or "RECOVERY".

Changed the first line of the message to be like the
Subject of the common notification mails. Especially having
the notification type at the beginning is great for reading
messages on the go ("RECOVERY? Don't even need to read the
remainder NOW").

Basicly removed the "*** monitoring" line. The only
relevant information in it was the monitoringhost.
I could argue that admins probably don't really need that
info at all (the affected host(displayname) should be more
unique, really), but nevertheless re-added the monitoring
host at the end in a smaller line for now.